requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
PyGithub==2.1.1

# Optional LLM providers - uncomment as needed
# anthropic==0.8.1
# google-generativeai==0.3.1
